server/requirements.txt
djangorestframework 
django==5.2
psycopg2-binary
python-dotenv 
django-environ
python-decouple>=3.8
djangorestframework-simplejwt>=5.3.0
django-jazzmin
django-cors-headers